+++ src/py42/clients/detectionlists/high_risk_employee.py
@@ -98,8 +98,8 @@ class HighRiskEmployeeClient(BaseClient):
         filter_type=None,
         sort_key=None,
         sort_direction=None,
-        page_num=None,
-        page_size=None,
+        page_num=1,
+        page_size=100,
     ):
         """Get a single page of High Risk Employees.
 
@@ -107,8 +107,8 @@ class HighRiskEmployeeClient(BaseClient):
             filter_type (str, optional): Valid filter types. Defaults to None.
             sort_key (str, optional): Sort results based by field. Defaults to None.
             sort_direction (str. optional): ``ASC`` or ``DESC``. Defaults to None.
-            page_num (str or int, optional): The page number to request. Defaults to None.
-            page_size (str or int, optional): The items to have per page. Defaults to None.
+            page_num (str or int, optional): The page number to request. Defaults to 1.
+            page_size (str or int, optional): The items to have per page. Defaults to 100.
 
         Returns:
             :class:`py42.response.Py42Response`